apps/control-plane-api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
argon2-cffi>=23.1.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
pytest>=7.0.0
